Expose same functionality in MaterialTextBox as TextBox
Because the MaterialTextBox doesn't extend the GWT standard TextBox, but
we still want to provide the same possibilities as the TextBox has, we
need to implement all the interfaces that TextBox does and redirect the
methods to the internal TextBox.

Also, the inner components should be protected instead of package
private, to expose them to extensions.