#192 MaterialSwitch problem with multiple ClickHandler registrations and ClickHandler order of registration
Move to Constructor in order to guarantee that it is only once
registered.
As well this resolved the problem that the value of the Action isn't
right when reading it in an onClick Handler. If the clickHandler is
registered in the constructor it is registered as first and therefore as
well always called as first what changes the value. So every later
called onClick handler in the chain will read the right value.